#235 build visualization fix
Builds can now again be filtered via the user list in the universal settings.
Instead of the github/lab users, it uses the author of the commit that triggered the build.
Removed User filtering from the issues visualization since the users in the universal settings are the git committers, while the users in the issue objects are the github/lab users. These two groups do not necessarily match.