feat: implement initial graph
The first draft of the graph was implemented. This includes both barcharts with the nodes in the middle.

A lot of values in styling are hardcoded and the graph is not responsive. Both aspects remain to be corrected.

Furthermore, the hover-over for the nodes for more information has to be also implemented.

Potential for additional work would be the adjustment of the bars depending on the probability of the label.

Ref:#237
Time-spent: 4h0m